Prevented app from crashing
I'm still not sure why the app crashes where it does, but at least I know
why now; when searching for the restaurant through Bing Maps, the result
is an array of possible locations. I just take the first location and zoom
the map in on it. When this array is empty, attempting to
get the first in that array obviously results in a crash.

All the addresses we're querying for should exist, though, since they come
straight out of the Yelp query. Well, who knows.